solved gpu memory problem by using Mat::moveTo(), copy everything to new Mat, then release the former one.

+++ src/cu_matrix_maths.h
@@ -15,7 +15,7 @@ __global__ void cu_elementWiseMultiply(const float*, const float*, float*, const
 __global__ void cu_setAll(float*, const float, const int);
 __global__ void cu_exp(const float*, float*, const int);
 __global__ void cu_log(const float*, float*, const int);
-__global__ void cu_pow(const float*, float*, const int, const int);
+__global__ void cu_pow(const float*, float*, const float, const int);
 __global__ void cu_divide(const float*, float*, const float, const int);
 __global__ void cu_divide(const float, const float*, float*, const int);
 __global__ void cu_divide(const float*, const float*, float*, const int);
